# Conflicts:
#	compiler/fory_compiler/tests/test_generated_code.py
#	cpp/fory/serialization/context.cc
#	cpp/fory/serialization/struct_serializer.h
#	cpp/fory/serialization/struct_test.cc
#	cpp/fory/serialization/type_resolver.cc
#	cpp/fory/serialization/type_resolver.h
#	csharp/src/Fory/Attributes.cs
#	csharp/src/Fory/TypeMeta.cs
#	csharp/tests/Fory.Tests/ForyRuntimeTests.cs
#	dart/packages/fory/lib/src/codegen/fory_generator.dart
#	dart/packages/fory/lib/src/resolver/type_resolver.dart
#	dart/packages/fory/lib/src/util/hash_util.dart
#	dart/packages/fory/test/struct_inheritance_generator_test.dart
#	dart/packages/fory/test/xlang_protocol_test.dart
#	docs/specification/xlang_serialization_spec.md
#	go/fory/tag.go
#	go/fory/type_resolver.go
#	java/fory-annotation-processor/src/test/java/org/apache/fory/annotation/processing/ForyStructProcessorTest.java
#	java/fory-core/src/main/java/org/apache/fory/meta/NativeTypeDefDecoder.java
#	java/fory-core/src/main/java/org/apache/fory/meta/TypeDef.java
#	java/fory-core/src/main/java/org/apache/fory/meta/TypeDefDecoder.java
#	java/fory-core/src/test/java/org/apache/fory/annotation/ForyFieldTagIdTest.java
#	java/fory-core/src/test/java/org/apache/fory/meta/NativeTypeDefEncoderTest.java
#	java/fory-core/src/test/java/org/apache/fory/meta/TypeDefEncoderTest.java
#	javascript/packages/core/lib/meta/TypeMeta.ts
#	javascript/test/typemeta.test.ts
#	kotlin/fory-kotlin-ksp/src/test/kotlin/org/apache/fory/kotlin/ksp/ProcessorValidationTest.kt
#	python/pyfory/field.py
#	python/pyfory/meta/typedef.py
#	python/pyfory/tests/test_typedef_encoding.py
#	rust/fory-core/src/meta/type_meta.rs
#	rust/fory-derive/src/object/read.rs
#	swift/Sources/Fory/TypeMeta.swift
#	swift/Sources/ForyMacro/ForyObjectMacro.swift
#	swift/Tests/ForyTests/CompatibilityTests.swift
#	swift/Tests/ForyTests/ForySwiftTests.swift
#	swift/Tests/ForyTests/MacroDiagnosticTests.swift

+++ cpp/fory/serialization/type_resolver.h
@@ -202,7 +202,7 @@ bool field_types_compatible_top_level(const FieldType &local,
 class FieldInfo {
 public:
   int32_t field_id;         // Wire tag ID; -1 means name-based identity
-  int32_t matched_field_id; // Derived compatible dispatch; -1 means unmatched
+  int16_t matched_field_id; // Local compatible dispatch; -1 means unmatched
   std::string field_name;   // Field name
   FieldType field_type;     // Field type information
 
@@ -281,8 +281,7 @@ public:
   /// get sorted field infos (sorted according to xlang spec)
   static std::vector<FieldInfo> sort_field_infos(std::vector<FieldInfo> fields);
 
-  /// Assign field IDs by comparing with local type
-  /// This is the key function for schema evolution!
+  /// Assign local compatible-reader dispatch IDs without changing wire tags.
   static Result<void, Error>
   assign_local_dispatch_ids(const TypeMeta *local_type,
                             std::vector<FieldInfo> &remote_fields);
